Using power bi and power query getting the profiles from TMP PORTAL, split the profiles based on skills and location and separate the evaluated profiles from the new profile
In 2022, done automation to collect the TMP profiles from TMP portal and save the profiles in excel sheet using power bi. Collect the TMP profiles in card view and save as in HTML format. Using Power BI, do refresh and collect the sheets based on skills, locations, grades and availability for all profiles.

In 2023, we are maintaining KBC profile tracker with the status. On Weekly twice, we are doing TMP portal  profiles refresh to collect the new profiles using power BI. By comparing the KBC profile tracker with the TMP profile sheet, we are separating the evaluated profiles from the new profiles using power query fuzzy logic.(comparing the associate name or id)